datastore: break cloudstore and filestore out in files This was done because the file was getting too big and AI was struggling to work with it.
